+++ packages/rpmtw_api_client/lib/src/models/cosmic_chat/cosmic_chat_info.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 class CosmicChatInfo {
   /// Number of online users.
   final int onlineUsers;