+++ packages/rpmtw_api_client/lib/src/models/cosmic_chat/cosmic_chat_message.dart
@@ -16,7 +16,7 @@ class CosmicChatMessage extends BaseModel {
 
   final String avatarUrl;
 
-  /// message sent time
+  /// message sent time (UTC+0)
   final DateTime sentAt;
 
   final CosmicChatUserType userType;