+++ packages/rpmtw_api_client/lib/src/models/storage/storage.dart
@@ -77,8 +77,13 @@ class Storage implements BaseModel {
         createAt.hashCode;
   }
 
-  static String getDownloadUrl(String uuid, {String? baseUrl}) {
-    baseUrl ??= RPMTWApiClient.instance.apiBaseUrl;
-    return '$baseUrl/storage/$uuid/download';
+  /// Get download url of the storage
+  /// 
+  /// **Parameters**
+  /// * [uuid] uuid of the storage
+  /// * [apiBaseUrl] api base url (optional)
+  static String getDownloadUrl(String uuid, {String? apiBaseUrl}) {
+    apiBaseUrl ??= RPMTWApiClient.instance.apiBaseUrl;
+    return '$apiBaseUrl/storage/$uuid/download';
   }
 }